pyproject.toml
[tool.poetry]
name = "working-cif-service"
version = "0.1.0"
description = ""
authors = ["jerryliang <jerryliang@jerryliang.top>"]
readme = "README.md"
packages = [{include = "working_cif_service"}]

[tool.poetry.dependencies]
python = "^3.9"
pyqt6 = "^6.4.1"
fitz = "^0.0.1.dev2"
pymupdf = "^1.21.1"
tencentcloud-sdk-python = "^3.0.822"
imbox = {git = "https://github.com/martinrusev/imbox.git"}
sqlalchemy = "^2.0.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
